errors and reprompt done

+++ exec/exit.c
@@ -6,7 +6,7 @@
 /*   By: asemsey <asemsey@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/15 13:44:48 by fnikzad           #+#    #+#             */
-/*   Updated: 2024/04/19 14:30:20 by asemsey          ###   ########.fr       */
+/*   Updated: 2024/04/20 11:58:56 by asemsey          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,7 +57,7 @@ int	ft_exit(t_mini *shell)
 	if (arr_len(cmd->argv) == 2)
 	{
 		shell->exit_status = ft_atol(cmd->argv[1]);
-		if (shell->exit_status == LLONG_MAX || shell->exit_status == LLONG_MIN)
+		if (shell->exit_status == INT_MAX || shell->exit_status == INT_MIN)
 		{
 			ft_putendl_fd("exit: numeric argument required", 2);
 		}
